+++ radix-engine-tests/tests/reference.rs
@@ -1,9 +1,5 @@
-use radix_engine::{
-    errors::{CallFrameError, KernelError, RuntimeError},
-    kernel::call_frame::{CloseSubstateError, MoveModuleError},
-    types::*,
-};
 use radix_engine::errors::SystemError;
+use radix_engine::{errors::RuntimeError, types::*};
 use radix_engine_interface::blueprints::resource::FromPublicKey;
 use scrypto_unit::*;
 use transaction::prelude::*;
